requirements.txt
Django
django-environ

# only for restframework
djangorestframework
django-rest-swagger
django-cors-headers

mysqlclient
tweepy
numpy
pandas
twint
pytrends